container: move idmap stuff to its own package
Users of the "container" package now have to link against liblxc, and this
idmap code is useful on its own.
